# docs/api/OPENAPI_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: PwC Data Engineering Platform API
  description: |
    ## Enterprise-grade REST API for retail data analytics

    The PwC Data Engineering Platform API provides comprehensive access to retail analytics data, 
    advanced analytics capabilities, and business intelligence insights. Built with FastAPI, 
    this production-ready API includes:

    - **JWT Authentication** with role-based access control
    - **Rate Limiting** with intelligent burst protection  
    - **Circuit Breaker** patterns for resilience
    - **Multi-level Caching** for optimal performance
    - **Comprehensive Monitoring** with distributed tracing
    - **Data Quality** validation and scoring
    - **Advanced Analytics** with ML-powered insights

    ### Authentication
    This API uses JWT (JSON Web Tokens) for authentication. Include your token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ### Rate Limits
    - **Authentication endpoints**: 5 requests per minute
    - **Analytics endpoints**: 20 requests per minute (burst: 30)
    - **General endpoints**: 100 requests per minute (burst: 150)

    ### API Versions
    - **v1**: Stable CRUD operations and basic analytics
    - **v2**: Advanced analytics with ML capabilities
  version: 2.0.0
  termsOfService: https://pwc.com/terms
  contact:
    name: PwC Data Engineering Team
    email: data-engineering@pwc.com
    url: https://docs.pwc-data-platform.com
  license:
...
servers:
...
security:
  - BearerAuth: []
  - BasicAuth: []
...
paths:
...
  /api/v1/sales:
    get:
      tags: [Sales v1]
      summary: List sales transactions
      description: |
        Retrieve paginated sales transactions with comprehensive filtering options.
        Supports date range filtering, product/country filtering, and flexible sorting.
      parameters:
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - $ref: '#/components/parameters/Product'
        - $ref: '#/components/parameters/Country'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: List of sales transactions
          headers:
            X-Total-Count:
              description: Total number of records
              schema:
                type: integer
            X-Page-Count:
              description: Total number of pages
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSales'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /api/v1/sales/{transaction_id}:
    get:
      tags: [Sales v1]
      summary: Get sales transaction by ID
      description: Retrieve detailed information for a specific sales transaction
      parameters:
        - name: transaction_id
          in: path
          required: true
          description: Unique transaction identifier
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Sales transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleItem'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/v2/sales:
    get:
      tags: [Sales v2]
      summary: Enhanced sales data with ML features
      description: |
        Advanced sales data endpoint with machine learning enrichment, 
        customer segmentation, and predictive features.
      parameters:
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - $ref: '#/components/parameters/Product'
        - $ref: '#/components/parameters/Country'
        - name: include_ml_features
          in: query
          description: Include machine learning features in response
          schema:
            type: boolean
            default: false
        - name: customer_segment
          in: query
          description: Filter by customer segment
          schema:
            type: string
            enum: [new_customers, loyal_customers, at_risk, champions]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Enhanced sales data with ML features
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnhancedPaginatedSales'

  # Advanced Analytics v2 Endpoints
  /api/v2/analytics/advanced-analytics:
    post:
      tags: [Analytics v2]
      summary: Create advanced analytics request
      description: |
        Submit complex analytics request for asynchronous processing.
        Supports advanced metrics calculation, customer segmentation,
        and multi-dimensional analysis with ML insights.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvancedAnalyticsRequest'
            example:
              start_date: "2024-01-01T00:00:00Z"
              end_date: "2024-01-31T23:59:59Z"
              metrics: ["revenue", "customers", "retention_rate"]
              dimensions: ["country", "product_category", "time"]
              filters:
                country: ["United Kingdom", "Germany"]
                min_order_value: 10.0
      responses:
        '202':
          description: Analytics request accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResult'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v2/analytics/analytics-status/{request_id}:
    get:
      tags: [Analytics v2]
      summary: Get analytics processing status
      description: Check the status of an advanced analytics request
      parameters:
        - name: request_id
          in: path
          required: true
          description: Analytics request identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Analytics processing status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResult'
        '404':
          description: Analytics request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v2/analytics/cohort-analysis:
    post:
      tags: [Analytics v2]
      summary: Customer cohort analysis
      description: |
        Perform sophisticated customer cohort analysis to understand
        customer retention patterns and lifetime value progression.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CohortAnalysisRequest'
      responses:
        '200':
          description: Cohort analysis results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CohortAnalysisResponse'
  /api/v2/analytics/predictive-insights:
    get:
      tags: [Analytics v2]
      summary: Predictive analytics and forecasting
      description: |
        Generate ML-powered predictive insights including revenue forecasting,
        customer behavior prediction, and trend analysis.
      parameters:
        - name: forecast_days
          in: query
          description: Number of days to forecast
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 30
        - name: confidence_interval
          in: query
          description: Confidence interval for predictions
          schema:
            type: number
            format: float
            minimum: 0.8
            maximum: 0.99
            default: 0.95
        - name: include_seasonality
          in: query
          description: Include seasonal patterns in forecast
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Predictive insights and forecasts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictiveInsights'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/v1/auth/login

    BasicAuth:
      type: http
      scheme: basic
      description: Basic HTTP authentication (fallback for development)
  parameters:
    DateFrom:
      name: date_from
      in: query
      description: Start date filter (ISO 8601 format)
      schema:
        type: string
        format: date
        example: "2024-01-01"
    DateTo:
      name: date_to
      in: query
      description: End date filter (ISO 8601 format)
      schema:
        type: string
        format: date
        example: "2024-01-31"
    Product:
      name: product
      in: query
      description: Product code filter
      schema:
        type: string
        example: "85123A"
    Country:
      name: country
      in: query
      description: Country name filter
      schema:
        type: string
        example: "United Kingdom"
    Page:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    Size:
      name: size
      in: query
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 20
        example: 100
    Sort:
      name: sort
      in: query
      description: Sort field and direction (field:asc|desc)
      schema:
        type: string
        default: "invoice_date:desc"
        example: "total_amount:desc"
  schemas:
...
    SaleItem:
      type: object
      required:
        - invoice_no
        - stock_code
        - quantity
        - invoice_date
        - unit_price
        - total
      properties:
        transaction_id:
          type: string
          format: uuid
          description: Unique transaction identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        invoice_no:
          type: string
          description: Invoice number
          example: "536365"
        stock_code:
          type: string
          description: Product stock code
          example: "85123A"
        description:
          type: string
          nullable: true
          description: Product description
          example: "WHITE HANGING HEART T-LIGHT HOLDER"
        quantity:
          type: integer
          description: Quantity purchased
          example: 6
        invoice_date:
          type: string
          format: date-time
          description: Invoice date and time
          example: "2024-01-01T08:26:00Z"
        unit_price:
          type: number
          format: float
          description: Price per unit
          example: 2.55
        customer_id:
          type: string
          nullable: true
          description: Customer identifier
          example: "17850"
        country:
          type: string
          nullable: true
          description: Customer country
          example: "United Kingdom"
        total:
          type: number
          format: float
          description: Total amount (quantity × unit_price)
          example: 15.30
        total_str:
          type: string
          description: Formatted total amount
          example: "15.30"
        data_quality_score:
          type: number
          format: float
          description: Data quality score (0.0-1.0)
          example: 0.95
        created_at:
          type: string
          format: date-time
          description: Record creation timestamp
          example: "2024-01-01T08:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Record last update timestamp
          example: "2024-01-01T08:30:00Z"
    PaginatedSales:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleItem'
        total:
          type: integer
          description: Total number of records
          example: 1000
        page:
          type: integer
          description: Current page number
          example: 1
        size:
          type: integer
          description: Page size
          example: 20
        pages:
          type: integer
          description: Total number of pages
          example: 50
    EnhancedPaginatedSales:
      allOf:
        - $ref: '#/components/schemas/PaginatedSales'
        - type: object
          properties:
            items:
              type: array
              items:
                allOf:
                  - $ref: '#/components/schemas/SaleItem'
                  - type: object
                    properties:
                      ml_features:
                        type: object
                        description: Machine learning enriched features
                        properties:
                          customer_segment:
                            type: string
                            enum: [new_customers, loyal_customers, at_risk, champions]
                            example: "loyal_customers"
                          predicted_clv:
                            type: number
                            format: float
                            description: Predicted customer lifetime value
                            example: 1250.75
                          churn_probability:
                            type: number
                            format: float
                            description: Probability of customer churn (0.0-1.0)
                            example: 0.15
                          product_affinity:
                            type: array
                            items:
                              type: object
                              properties:
                                product_code:
                                  type: string
                                  example: "85123B"
                                affinity_score:
                                  type: number
                                  format: float
                                  example: 0.85
            aggregations:
              type: object
              description: Aggregated statistics for current filter
              properties:
                total_revenue:
                  type: number
                  format: float
                  example: 125000.50
                unique_customers:
                  type: integer
                  example: 8945
                avg_order_value:
                  type: number
                  format: float
                  example: 27.63

    # Analytics Schemas
...
    AdvancedAnalyticsRequest:
      type: object
      required:
        - start_date
        - end_date
      properties:
        start_date:
          type: string
          format: date-time
          description: Analysis start date
          example: "2024-01-01T00:00:00Z"
        end_date:
          type: string
          format: date-time
          description: Analysis end date
          example: "2024-01-31T23:59:59Z"
        metrics:
          type: array
          items:
            type: string
            enum: [revenue, transactions, customers, retention_rate, clv, churn_rate]
          default: [revenue, transactions]
          example: [revenue, customers, retention_rate]
        dimensions:
          type: array
          items:
            type: string
            enum: [country, product, time, customer_segment, product_category]
          default: [country, product]
          example: [country, product_category, time]
        filters:
          type: object
          nullable: true
          description: Additional filters to apply
          example:
            country: ["United Kingdom", "Germany"]
            min_order_value: 10.0
            customer_segment: ["loyal_customers", "champions"]
    AnalyticsResult:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
          description: Unique request identifier
          example: "req_550e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          enum: [processing, completed, failed]
          description: Processing status
          example: "completed"
        data:
          type: object
          nullable: true
          description: Analytics results (null if still processing)
          properties:
            revenue_analysis:
              type: object
              properties:
                total_revenue:
                  type: number
                  format: float
                  example: 1250000.50
                growth_rate:
                  type: number
                  format: float
                  example: 15.2
                top_products:
                  type: array
                  items:
                    type: string
                  example: ["Product A", "Product B", "Product C"]
            customer_analysis:
              type: object
              properties:
                total_customers:
                  type: integer
                  example: 45230
                new_customers:
                  type: integer
                  example: 1250
                retention_rate:
                  type: number
                  format: float
                  example: 87.5
        metadata:
          type: object
          properties:
            processing_time:
              type: string
              example: "45.2 seconds"
            estimated_duration:
              type: string
              example: "30-60 seconds"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        completed_at:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-15T10:30:45Z"
    CohortAnalysisRequest:
      type: object
      required:
        - start_date
        - end_date
      properties:
        start_date:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        end_date:
          type: string
          format: date-time
          example: "2024-01-31T23:59:59Z"
        cohort_period:
          type: string
          enum: [day, week, month]
          default: month
          description: Period for cohort grouping
        analysis_period:
          type: string
          enum: [day, week, month]
          default: month
          description: Period for retention analysis
    CohortAnalysisResponse:
      type: object
      properties:
        cohort_periods:
          type: array
          items:
            type: string
          example: ["2024-01", "2024-02", "2024-03"]
        retention_rates:
          type: object
          additionalProperties:
            type: number
            format: float
          example:
            month_0: 100.0
            month_1: 45.2
            month_2: 32.1
            month_3: 28.5
        cohort_sizes:
          type: array
          items:
            type: integer
          example: [1200, 1350, 1180]
        revenue_per_cohort:
          type: array
          items:
            type: object
            properties:
              period:
                type: string
                example: "2024-01"
              revenue:
                type: number
                format: float
                example: 45000.0
    PredictiveInsights:
      type: object
      properties:
        forecast_period:
          type: integer
          description: Number of days forecasted
          example: 30
        confidence_interval:
          type: number
          format: float
          description: Confidence interval used
          example: 0.95
        predictions:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date-time
                example: "2024-02-01T00:00:00Z"
              predicted_revenue:
                type: number
                format: float
                example: 12500.50
              lower_bound:
                type: number
                format: float
                example: 11250.45
              upper_bound:
                type: number
                format: float
                example: 13750.55
        model_accuracy:
          type: number
          format: float
          description: Model accuracy score
          example: 0.85
        last_updated:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Request validation failed"
            details:
              type: object
              description: Additional error details
              properties:
                field:
                  type: string
                  example: "customer_id"
                message:
                  type: string
                  example: "Field is required"
            timestamp:
              type: string
              format: date-time
              example: "2024-01-15T10:30:00Z"
            request_id:
              type: string
              description: Unique request identifier for troubleshooting
              example: "req_abc123"
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
              timestamp: "2024-01-15T10:30:00Z"
              request_id: "req_abc123"
...
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Time when rate limit resets (Unix timestamp)
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retry
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "RATE_LIMIT_EXCEEDED"
              message: "Rate limit exceeded. Please retry after 60 seconds."
              timestamp: "2024-01-15T10:30:00Z"
              request_id: "req_abc123"
...