# src/api/v2/routes/analytics.py
"""
Advanced Analytics API v2
Enterprise-grade analytics endpoints with caching and async processing
"""
from datetime import datetime, timedelta
from typing import Any

from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, Query
from pydantic import BaseModel, Field

from core.caching.redis_cache import RedisCache
from core.logging import get_logger
from domain.services.sales_service import SalesService

logger = get_logger(__name__)
router = APIRouter(prefix="/analytics", tags=["Analytics v2"])


class AnalyticsRequest(BaseModel):
    start_date: datetime = Field(..., description="Start date for analysis")
    end_date: datetime = Field(..., description="End date for analysis")
    metrics: list[str] = Field(default=["revenue", "transactions"], description="Metrics to calculate")
    dimensions: list[str] = Field(default=["country", "product"], description="Dimensions to group by")
    filters: dict[str, Any] | None = Field(default=None, description="Additional filters")


class AnalyticsResult(BaseModel):
    request_id: str = Field(..., description="Unique request identifier")
    status: str = Field(..., description="Processing status")
    data: dict[str, Any] | None = Field(default=None, description="Analytics results")
    metadata: dict[str, Any] = Field(default_factory=dict, description="Processing metadata")
    created_at: datetime = Field(default_factory=datetime.utcnow)
    completed_at: datetime | None = Field(default=None)


class CohortAnalysisRequest(BaseModel):
    start_date: datetime
    end_date: datetime
    cohort_period: str = Field(default="month", description="Cohort period: day, week, month")
    analysis_period: str = Field(default="month", description="Analysis period: day, week, month")


@router.post("/advanced-analytics", response_model=AnalyticsResult)
async def create_advanced_analytics(
    request: AnalyticsRequest,
    background_tasks: BackgroundTasks,
    cache: RedisCache = Depends(),
    sales_service: SalesService = Depends()
):
    """
    Create advanced analytics calculation with async processing
    """
    import uuid
    request_id = str(uuid.uuid4())

    # Check cache first
    cache_key = f"analytics:{hash(str(request.dict()))}"
    cached_result = await cache.get(cache_key)

    if cached_result:
        logger.info(f"Returning cached analytics result for request {request_id}")
        return AnalyticsResult(
            request_id=request_id,
            status="completed",
            data=cached_result,
            metadata={"source": "cache"}
        )

    # Queue background processing
    background_tasks.add_task(
        process_advanced_analytics,
        request_id=request_id,
        analytics_request=request,
        cache_key=cache_key
    )

    # Store initial result
    initial_result = AnalyticsResult(
        request_id=request_id,
        status="processing",
        metadata={"estimated_duration": "30-60 seconds"}
    )

    await cache.set(f"analytics_status:{request_id}", initial_result.dict(), ttl=3600)

    return initial_result


@router.get("/analytics-status/{request_id}", response_model=AnalyticsResult)
async def get_analytics_status(
    request_id: str,
    cache: RedisCache = Depends()
):
    """
    Get status of analytics processing
    """
    status_data = await cache.get(f"analytics_status:{request_id}")

    if not status_data:
        raise HTTPException(status_code=404, detail="Analytics request not found")

    return AnalyticsResult(**status_data)


@router.post("/cohort-analysis", response_model=dict[str, Any])
async def cohort_analysis(
    request: CohortAnalysisRequest,
    cache: RedisCache = Depends(),
    sales_service: SalesService = Depends()
):
    """
    Perform customer cohort analysis
    """
    cache_key = f"cohort:{hash(str(request.dict()))}"
    cached_result = await cache.get(cache_key)

    if cached_result:
        return cached_result

    try:
        # Simulate cohort analysis calculation
        cohort_data = await calculate_cohort_analysis(request, sales_service)

        # Cache for 1 hour
        await cache.set(cache_key, cohort_data, ttl=3600)

        return cohort_data

    except Exception as e:
        logger.error(f"Cohort analysis failed: {e}")
        raise HTTPException(status_code=500, detail="Cohort analysis failed")


@router.get("/predictive-insights", response_model=dict[str, Any])
async def predictive_insights(
    forecast_days: int = Query(default=30, ge=1, le=365),
    confidence_interval: float = Query(default=0.95, ge=0.8, le=0.99),
    cache: RedisCache = Depends()
):
    """
    Generate predictive insights and forecasting
    """
    cache_key = f"forecast:{forecast_days}:{confidence_interval}"
    cached_result = await cache.get(cache_key)

    if cached_result:
        return cached_result

    # Simulate ML model predictions
    predictions = await generate_predictions(forecast_days, confidence_interval)

    # Cache for 4 hours
    await cache.set(cache_key, predictions, ttl=14400)

    return predictions


@router.get("/real-time-metrics", response_model=dict[str, Any])
async def real_time_metrics(
    metrics: list[str] = Query(default=["active_users", "revenue_rate", "conversion_rate"]),
    cache: RedisCache = Depends()
):
    """
    Get real-time business metrics
    """
    cache_key = f"realtime:{':'.join(sorted(metrics))}"
    cached_result = await cache.get(cache_key)

    if cached_result:
        return cached_result

    # Calculate real-time metrics
    real_time_data = await calculate_real_time_metrics(metrics)

    # Cache for 30 seconds
    await cache.set(cache_key, real_time_data, ttl=30)

    return real_time_data


async def process_advanced_analytics(
    request_id: str,
    analytics_request: AnalyticsRequest,
    cache_key: str
):
    """
    Background task to process complex analytics
    """
    try:
        # Simulate complex analytics processing
        await asyncio.sleep(2)  # Simulate processing time

        result_data = {
            "revenue_analysis": {
                "total_revenue": 1250000.50,
                "growth_rate": 15.2,
                "top_products": ["Product A", "Product B", "Product C"]
            },
            "customer_analysis": {
                "total_customers": 45230,
                "new_customers": 1250,
                "retention_rate": 87.5
            },
            "geographic_analysis": {
                "top_countries": ["United Kingdom", "Germany", "France"],
                "revenue_by_country": {
                    "United Kingdom": 650000.25,
                    "Germany": 320000.75,
                    "France": 280000.50
                }
            }
        }

        # Update status
        from core.caching.redis_cache import RedisCache
        cache = RedisCache()

        completed_result = AnalyticsResult(
            request_id=request_id,
            status="completed",
            data=result_data,
            metadata={"processing_time": "2.1 seconds"},
            completed_at=datetime.utcnow()
        )

        await cache.set(f"analytics_status:{request_id}", completed_result.dict(), ttl=3600)
        await cache.set(cache_key, result_data, ttl=1800)  # Cache for 30 minutes

        logger.info(f"Completed advanced analytics processing for request {request_id}")

    except Exception as e:
        logger.error(f"Analytics processing failed for request {request_id}: {e}")


async def calculate_cohort_analysis(
    request: CohortAnalysisRequest,
    sales_service: SalesService
) -> dict[str, Any]:
    """
    Calculate customer cohort analysis
    """
    # Simulate cohort calculation
    return {
        "cohort_periods": ["2024-01", "2024-02", "2024-03", "2024-04"],
        "retention_rates": {
            "month_0": 100.0,
            "month_1": 45.2,
            "month_2": 32.1,
            "month_3": 28.5,
            "month_4": 25.8
        },
        "cohort_sizes": [1200, 1350, 1180, 1420],
        "revenue_per_cohort": [
            {"period": "2024-01", "revenue": 45000.0},
            {"period": "2024-02", "revenue": 52000.0},
            {"period": "2024-03", "revenue": 48000.0},
            {"period": "2024-04", "revenue": 58000.0}
        ]
    }


async def generate_predictions(forecast_days: int, confidence_interval: float) -> dict[str, Any]:
    """
    Generate ML-based predictions
    """
    import random
    base_revenue = 10000

    predictions = []
    for i in range(forecast_days):
        trend = i * 50  # Upward trend
        noise = random.uniform(-500, 500)  # Random variation
        predicted_value = base_revenue + trend + noise

        ci_range = predicted_value * (1 - confidence_interval) * 0.1

        predictions.append({
            "date": (datetime.utcnow() + timedelta(days=i+1)).isoformat(),
            "predicted_revenue": round(predicted_value, 2),
            "lower_bound": round(predicted_value - ci_range, 2),
            "upper_bound": round(predicted_value + ci_range, 2)
        })

    return {
        "forecast_period": forecast_days,
        "confidence_interval": confidence_interval,
        "predictions": predictions,
        "model_accuracy": 0.85,
        "last_updated": datetime.utcnow().isoformat()
    }


async def calculate_real_time_metrics(metrics: list[str]) -> dict[str, Any]:
    """
    Calculate real-time business metrics
    """
    import random

    metric_data = {}

    for metric in metrics:
        if metric == "active_users":
            metric_data[metric] = {
                "value": random.randint(450, 650),
                "change_24h": random.uniform(-5.0, 12.0),
                "unit": "users"
            }
        elif metric == "revenue_rate":
            metric_data[metric] = {
                "value": random.uniform(800.0, 1200.0),
                "change_24h": random.uniform(-8.0, 15.0),
                "unit": "$/hour"
            }
        elif metric == "conversion_rate":
            metric_data[metric] = {
                "value": random.uniform(2.5, 4.2),
                "change_24h": random.uniform(-0.5, 0.8),
                "unit": "%"
            }

    return {
        "metrics": metric_data,
        "timestamp": datetime.utcnow().isoformat(),
        "refresh_interval": 30
    }